requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.4.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.45
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
